% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsef.R
\name{collapsef}
\alias{collapsef}
\title{Collapse vector to string.}
\usage{
collapsef(..., sep = ",")
}
\arguments{
\item{...}{[any]\cr
See \code{\link{sprintf}}.}

\item{sep}{[\code{character(1)}]\cr
See \code{\link{collapse}}.}
}
\value{
[\code{character(1)}].
}
\description{
A simple wrapper for \code{collapse(sprintf, ...)}.
}
\details{
Useful for vectorized call to \code{\link{sprintf}}.
}

