% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ensureLength}
\alias{ensureLength}
\title{Ensure that an input is a vector / list of a certain length.}
\usage{
ensureLength(x, n, names = NULL)
}
\arguments{
\item{x}{[any]\cr
Input element.}

\item{n}{[\code{integer}]\cr
Desired length.}

\item{names}{[\code{character}]\cr
Names for result.
Default is NULL, which means no names.}
}
\value{
[\code{vector} | \code{list}].
}
\description{
Useful for standard argument conversion where a user can input a single
element, but this has to be replicated now n times for a resulting vector or list.
}

