\name{explode}
\alias{explode}
\title{Split a string into substrings.}
\usage{
explode(x, sep = " ")
}
\arguments{
  \item{x}{[\code{character}]\cr Source string.}

  \item{sep}{[\code{character}]\cr Seperator whcih is used
  to split \code{x} into substrings.  Default is \dQuote{
  }.}
}
\value{
[\code{vector}] Vector of substrings.
}
\description{
Split up a string into substrings according to a seperator.
}

