% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterNull.R
\name{filterNull}
\alias{filterNull}
\title{Filter a list for NULL values}
\usage{
filterNull(li)
}
\arguments{
\item{li}{[\code{list}]\cr
List.}
}
\value{
[\code{list}].
}
\description{
Filter a list for NULL values
}
