% Generated by roxygen2 (4.0.1): do not edit by hand
<<<<<<< HEAD:man/getClass.Rd
\name{getClass}
\alias{getClass}
\title{Wrapper for \code{class(x)[1]}.}
||||||| merged common ancestors
\name{getClass}
\alias{getClass}
\title{Wrapper for \code{class(x)[1L]}.}
=======
\name{getClass1}
\alias{getClass1}
\title{Wrapper for \code{class(x)[1L]}.}
>>>>>>> master:man/getClass1.Rd
\usage{
getClass1(x)
}
\arguments{
\item{x}{[any]\cr
Input object.}
}
\value{
[\code{character(1)}].
}
\description{
Wrapper for \code{class(x)[1]}.
}
\note{
\code{getClass} is a function in \code{methods}. Do not confuse.
}

