% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasAttributes.R
\name{hasAttributes}
\alias{hasAttributes}
\title{Check if given object has certain attributes}
\usage{
hasAttributes(obj, attribute.names)
}
\arguments{
\item{obj}{[mixed]\cr
Arbitrary R object.}

\item{attribute.names}{[\code{character}]\cr
Vector of strings, i.e., attribute names.}
}
\value{
[\code{logical(1)}]
  \code{TRUE} if object \code{x} contains all attributes from \code{attributeNames}
  and \code{FALSE} otherwise.
}
\description{
Checks whether an object has all specified attributes.
}
