% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFALSE.R
\name{isFALSE}
\alias{isFALSE}
\title{A wrapper for \code{identical(x, FALSE)}}
\usage{
isFALSE(x)
}
\arguments{
\item{x}{[any]\cr
Your object.}
}
\value{
[\code{logical(1)}].
}
\description{
Tests if an object is identical to FALSE.
}
\examples{
isFALSE(0)
isFALSE(FALSE)
}
