% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isScalarNA.R
\name{isScalarNA}
\alias{isScalarNA}
\title{Checks whether an object is a scalar NA value.}
\usage{
isScalarNA(x)
}
\arguments{
\item{x}{[any]\cr
Object to check.}
}
\value{
[\code{logical(1)}].
}
\description{
Checks whether object is from \code{(NA, NA_integer, NA_real_, NA_character_, NA_complex_)}.
}
