% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedLapply.R
\name{namedLapply}
\alias{namedLapply}
\title{Like \code{lapply} but also passes the name to the function.}
\usage{
namedLapply(x, fun, ...)
}
\arguments{
\item{x}{[\code{list}]\cr
Named list.}

\item{fun}{[\code{function}]\cr
The function to apply. Second argument will be the name of the element.}

\item{...}{[any]\cr
Further arguments passed down to \code{fun}.}
}
\value{
[\code{list}].
}

