% Generated by roxygen2 (4.0.0): do not edit by hand
\name{\%nin\%}
\alias{\%nin\%}
\title{Simply a negated \code{in} operator.}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{[\code{vector}]\cr
Values that should not be in \code{y}.}

\item{y}{[\code{vector}]\cr
Values to match against.}
}
\description{
Simply a negated \code{in} operator.
}

