\name{parallelGetExported}
\alias{parallelGetExported}
\title{Retrieve a with \code{\link{parallelExport}} exported in slave code.}
\usage{
  parallelGetExported(name)
}
\arguments{
  \item{name}{[\code{character(1)}]\cr Name of exported
  object.}
}
\value{
  [any]. Object value.
}
\description{
  Retrieve a with \code{\link{parallelExport}} exported in
  slave code.
}

