% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printHead.R
\name{printHead}
\alias{printHead}
\title{More meaningful \code{head(df)} output.}
\usage{
printHead(x, n = 6L)
}
\arguments{
\item{x}{[\code{data.frame} | \code{matrix} | \code{list} | \code{vector}]\cr
Object.}

\item{n}{[\code{integer(1)}]\cr
Single positive integer: number of rows for a matrix/data.frame or number of
elements for vectors/lists respectively.}
}
\value{
Nothing.
}
\description{
The behaviour is similar to \code{print(head(x, n))}. The difference is, that if
the number of rows in a data.frame/matrix or the number of elements in a list
or vector is larger than \code{n}, additional information is printed about
the total number of rows or elements respectively.
}
