% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttribute.R
\name{setAttribute}
\alias{setAttribute}
\title{A wrapper for \code{attr(x, which) = y}}
\usage{
setAttribute(x, which, value)
}
\arguments{
\item{x}{[any]\cr
Your object.}

\item{which}{[\code{character(1)}]\cr
Name of the attribute to set}

\item{value}{[\code{ANY}]\cr
Value for the attribute.}
}
\value{
Changed object \code{x}.
}
\description{
Sets an attribute on an object and returns the modified object.
}
\examples{
setAttribute(list(), "foo", 1)
}
