% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitPath.R
\name{splitPath}
\alias{splitPath}
\title{Split a path into components}
\usage{
splitPath(path)
}
\arguments{
\item{path}{[\code{character(1)}]\cr
Path to split as string}
}
\value{
\code{named list}: List with components \dQuote{drive} (\code{character(1)}
 and \dQuote{path} (\code{character(n)}.
}
\description{
The first normalized path is split on forward and backward slashes and its components returned as
character vector. The drive or network home are extracted separately on windows systems and
empty on all other systems.
}

