% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressAll.R
\name{suppressAll}
\alias{suppressAll}
\title{Suppresses all output except for errors.}
\usage{
suppressAll(expr)
}
\arguments{
\item{expr}{[valid R expression]\cr
Expression.}
}
\value{
Return value of expression invisibly.
}
\description{
Evaluates an expression and suppresses all output except for errors, 
meaning: prints, messages, warnings and package startup messages.
}
\examples{
suppressAll({
  print("foo")
  message("foo")
  warning("foo")
})
}

