\name{usedFactorLevels}
\alias{usedFactorLevels}
\title{Determines used factor levels.}
\usage{
usedFactorLevels(x)
}
\arguments{
  \item{x}{[\code{factor}]\cr Source vector.}
}
\value{
[\code{character}] Character vector of used levels.
}
\description{
Determines the factor levels of a factor type vector that
are actually occuring in it.
}

