% Generated by roxygen2 (4.0.0): do not edit by hand
\name{addClasses}
\alias{addClasses}
\title{A wrapper to add to the class attribute.}
\usage{
addClasses(x, classes)
}
\arguments{
\item{x}{[any]\cr
Your object.}

\item{classes}{[\code{character}]\cr
Classes to add. Will be added in front (specialization).}
}
\value{
Changed object \code{x}.
}
\description{
A wrapper to add to the class attribute.
}
\examples{
addClasses(list(), c("foo1", "foo2"))
}

