% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse vector to string.}
\usage{
collapse(x, sep = ",")
}
\arguments{
\item{x}{[\code{vector}]\cr
Vector to collapse.}

\item{sep}{[\code{character(1)}]\cr
Passed to \code{collapse} in \code{\link{paste}}.
Default is \dQuote{,}.}
}
\value{
[\code{character(1)}].
}
\description{
A simple wrapper for \code{paste(x, collapse)}.
}
\examples{
collapse(c("foo", "bar"))
collapse(c("foo", "bar"), sep = ";")
}
