% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convertDataFrameCols}
\alias{convertDataFrameCols}
\title{Converts columns in a data frame to characters, factors or numerics.}
\usage{
convertDataFrameCols(df, chars.as.factor = FALSE, factors.as.char = FALSE,
  ints.as.num = FALSE, logicals.as.factor = FALSE)
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data frame.}

\item{chars.as.factor}{[\code{logical(1)}]\cr
Should characters be converted to factors?
Default is \code{FALSE}.}

\item{factors.as.char}{[\code{logical(1)}]\cr
Should characters be converted to factors?
Default is \code{FALSE}.}

\item{ints.as.num}{[\code{logical(1)}]\cr
Should integers be converted to numerics?
Default is \code{FALSE}.}

\item{logicals.as.factor}{[\code{logical(1)}]\cr
Should logicals be converted to factors?
Default is \code{FALSE}.}
}
\value{
[\code{data.frame}].
}
\description{
Converts columns in a data frame to characters, factors or numerics.
}

