% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertInteger}
\alias{convertInteger}
\title{Conversion for single integer.}
\usage{
convertInteger(x)
}
\arguments{
\item{x}{[any]\cr
Argument.}
}
\value{
Either a single integer if conversion was done or \code{x} unchanged.
}
\description{
Convert single numeric to integer only if the numeric represents a single integer,
e.g. 1 to 1L.
Otherwise the argument is returned unchanged.
}
\examples{
str(convertInteger(1.0))
str(convertInteger(1.3))
str(convertInteger(c(1.0, 2.0)))
str(convertInteger("foo"))
}

