% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropNamed.R
\name{dropNamed}
\alias{dropNamed}
\title{Drop named elements of an object.}
\usage{
dropNamed(x, drop = character(0L))
}
\arguments{
\item{x}{[any]\cr
Object to drop named elements from.
For a matrix or a data frames this function drops named columns via
the second argument of the binary index operator \code{[,]}.
Otherwise, the unary index operator \code{[]} is used for dropping.}

\item{drop}{[\code{character}]\cr
Names of elements to drop.}
}
\value{
Subset of object of same type as \code{x}. The object is not simplified,
  i.e, no dimensions are dropped as \code{[,,drop = FALSE]} is used.
}
\description{
Drop named elements of an object.
}
