% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ensureVector}
\alias{ensureVector}
\title{Blow up single scalars / objects to vectors /  list by replication.}
\usage{
ensureVector(x, n, cl = NULL, names = NULL)
}
\arguments{
\item{x}{[any]\cr
Input element.}

\item{n}{[\code{integer}]\cr
Desired length.}

\item{cl}{[\code{character(1)}*]\cr
Only do the operation if \code{x} inherits from this class, otherwise simply let x pass.
Default is \code{NULL} which means to always do the operation.}

\item{names}{[\code{character}*] \cr
Names for result.
Default is \code{NULL}, which means no names.}
}
\value{
Ether a vector or list of length \code{n} with replicated \code{x} or \code{x} unchanged..
}
\description{
Useful for standard argument conversion where a user can input a single
element, but this has to be replicated now n times for a resulting vector or list.
}

