% Generated by roxygen2 (4.0.2): do not edit by hand
\name{explode}
\alias{explode}
\title{Split up a string into substrings.}
\usage{
explode(x, sep = " ")
}
\arguments{
\item{x}{[\code{character}]\cr
Source string.}

\item{sep}{[\code{character}]\cr
Seperator whcih is used to split \code{x} into substrings.
Default is \dQuote{ }.}
}
\value{
[\code{vector}]
  Vector of substrings.
}
\description{
Split up a string into substrings according to a seperator.
}
\examples{
explode("foo bar")
explode("comma,seperated,values", sep = ",")
}

