% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getAttributeNames}
\alias{getAttributeNames}
\title{Helper function for determining the vector of attribute names
of a given object.}
\usage{
getAttributeNames(obj)
}
\arguments{
\item{obj}{[any]\cr
Source object.}
}
\value{
[\code{character}]
  Vector of attribute names for the source object.
}
\description{
Helper function for determining the vector of attribute names
of a given object.
}

