% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClass1.R
\name{getClass1}
\alias{getClass1}
\title{Wrapper for \code{class(x)[1]}}
\usage{
getClass1(x)
}
\arguments{
\item{x}{[any]\cr
Input object.}
}
\value{
[\code{character(1)}].
}
\description{
Returns the first class of an object.
}
\note{
\code{getClass} is a function in \code{methods}. Do not confuse.
}
