% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getClass1.R
\name{getClass1}
\alias{getClass1}
\title{Wrapper for \code{class(x)[1]}.}
\usage{
getClass1(x)
}
\arguments{
\item{x}{[any]\cr
Input object.}
}
\value{
[\code{character(1)}].
}
\description{
Wrapper for \code{class(x)[1]}.
}
\note{
\code{getClass} is a function in \code{methods}. Do not confuse.
}

