% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getClass1}
\alias{getClass1}
\title{Wrapper for \code{class(x)[1L]}.}
\usage{
getClass1(x)
}
\arguments{
\item{x}{[any]\cr
Input object.}
}
\value{
[\code{character(1)}].
}
\description{
Wrapper for \code{class(x)[1L]}.
}
\note{
\code{getClass} is a function in \code{methods}. Do not confuse.
}

