% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstLast.R
\name{getFirst}
\alias{getFirst}
\alias{getLast}
\title{Get the first/last element of a list/vector}
\usage{
getFirst(x)

getLast(x)
}
\arguments{
\item{x}{[\code{list} | \code{vector}]\cr
The list or vector.}
}
\value{
Selected element. The element name is dropped.
}
\description{
Returns the first or last element of a list or vector.
}
