% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOperatingSystem.R
\name{getOperatingSystem}
\alias{getOperatingSystem}
\alias{isDarwin}
\alias{isLinux}
\alias{isUnix}
\alias{isWindows}
\title{Functions to determine the operating system.}
\usage{
getOperatingSystem()

isWindows()

isUnix()

isLinux()

isDarwin()
}
\value{
See above.
}
\description{
\itemize{
\item{getOperatingSystem}{Simple wrapper for \code{.Platform$OS.type}, returns \code{character(1)}.}
\item{isUnix}{Predicate for OS string, returns \code{logical(1)}. Currently this would include Unix, Linux and Mac flavours.}
\item{isLinux}{Predicate for sysname string, returns \code{logical(1)}.}
\item{isDarwin}{Predicate for sysname string, returns \code{logical(1)}.}
\item{isWindows}{Predicate for OS string, returns \code{logical(1)}.}
}
}

