% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getUsedFactorLevels}
\alias{getUsedFactorLevels}
\title{Determines used factor levels.}
\usage{
getUsedFactorLevels(x)
}
\arguments{
\item{x}{[\code{factor}]\cr
Source vector.}
}
\value{
[\code{character}]
  Character vector of used levels.
}
\description{
Determines the factor levels of a factor type vector
that are actually occuring in it.
}

