% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsedFactorLevels.R
\name{getUsedFactorLevels}
\alias{getUsedFactorLevels}
\title{Determines used factor levels}
\usage{
getUsedFactorLevels(x)
}
\arguments{
\item{x}{[\code{factor}]\cr
The factor.}
}
\value{
[\code{character}]
}
\description{
Determines the factor levels of a factor type vector
that are actually occuring in it.
}
