\name{isScalarValue}
\alias{isScalarValue}
\title{Is given argument an atomic vector or factor of length 1?}
\usage{
  isScalarValue(x, na.ok = TRUE, null.ok = FALSE)
}
\arguments{
  \item{x}{[any]\cr Argument.}

  \item{na.ok}{[\code{logical(1)}]\cr Is \code{NA}
  considered a scalar?  Default is \code{TRUE}.}

  \item{null.ok}{[\code{logical(1)}]\cr Is \code{NULL}
  considered a scalar?  Default is \code{FALSE}.}
}
\value{
  [\code{logical(1)}].
}
\description{
  Is given argument an atomic vector or factor of length 1?
}

