% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsort.R
\name{lsort}
\alias{lsort}
\title{A wrapper for \code{\link{sort}} to sort using the \dQuote{C} collating rules.}
\usage{
lsort(...)
}
\arguments{
\item{...}{Options passed to sort.}
}
\value{
See \code{\link{sort}}.
}
\description{
A wrapper for \code{\link{sort}} to sort using the \dQuote{C} collating rules.
}

