% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messagef.R
\name{messagef}
\alias{messagef}
\title{Wrapper for message and sprintf.}
\usage{
messagef(..., .newline = TRUE)
}
\arguments{
\item{...}{[any]\cr
See \code{\link{sprintf}}.}

\item{.newline}{[logical(1)]\cr
Add a newline to the message. Default is \code{TRUE}.}
}
\value{
Nothing.
}
\description{
A simple wrapper for \code{message(sprintf(...))}.
}
\examples{
msg = "a message"
warningf("this is \%s", msg)
}

