% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedList.R
\name{namedList}
\alias{namedList}
\title{Create named list, possibly initialized with a certain element.}
\usage{
namedList(names, init)
}
\arguments{
\item{names}{[\code{character}]\cr
Names of elements.}

\item{init}{[valid R expression]\cr
If given all list elements are initialized to this, otherwise
\code{NULL} is used.}
}
\value{
[\code{list}].
}
\description{
Even an empty list will always be named.
}
\examples{
namedList(c("a", "b"))
namedList(c("a", "b"), init = 1)
}
